chore: move test-cli files
