+++ test/commands/deploy.test.js
@@ -1,4 +1,4 @@
-const testCli = require('./test-cli');
+const testCli = require('./test-cli-helper/test-cli');
 const DeployCommand = require('../../src/commands/deploy');
 const { enter } = require('../fixtures/std');
 const {
