+++ test/commands/init.test.js
@@ -1,4 +1,4 @@
-const testCli = require('./test-cli');
+const testCli = require('./test-cli-helper/test-cli');
 const InitCommand = require('../../src/commands/init');
 const {
   getProjectByEnvIncludeLegacy,
