+++ test/commands/test-cli-helper/test-cli-api.js
