+++ test/commands/test-cli-helper/test-cli-auth-token.js
