+++ test/commands/test-cli-helper/test-cli-command.js
