+++ test/commands/test-cli-helper/test-cli-fs.js
