+++ test/commands/test-cli-helper/test-cli-std.js
