+++ test/commands/test-cli-helper/test-cli.js
